requirements.txt
comtypes==1.4.10
psutil==7.0.0
pycaw==20240210
deep-translator==1.11.4
PyQt5==5.15.11
PyQt5_sip==12.17.0
python-dotenv==1.0.1
rapidfuzz==3.12.2
SpeechRecognition==3.10.4
PyAudio==0.2.14
cffi==1.17.1
omegaconf==2.3.0
edge_tts==7.0.0
sounddevice==0.5.1
librosa==0.11.0
pillow==11.1.0
matplotlib==3.10.1
screen_brightness_control==0.24.2
cohere==5.14.0
dotenv==0.9.9
openai==1.72.0
spacy==3.8.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
pywin32==310
PyGithub==2.6.1
